Update Source Code
In this update I have discovered how to use the return values at the end of each operation of the function. I have also learned how to use recursion. In this update I have clear recursion, however, I have not succeeded in printing the value of the addition of each multiple of factors until the initial input reaches 1. So far I have taken the input down to 1 or 0 through use of recursion, but I have not succeeded in adding that MEANINGFUL operation which would update throughout, such as the sum of each multiplication for factor.