Discoveries on Recursive Properties
I've learned that a recursive algorithm should not be called more than once in the same function. A function that calls itself is synonymous with a terminated operation, its base case its only heed. I learned this through the fact that, when the FOR LOOP sets in motion, the counter, i, never augments beyond 1. This phenomenon of the counter not augmenting I am calling a "RECURSIVE PROPERTY," that is, repetition without evidence.

And so, in order to get the sum of each recursed integer, I must utilize pointers. But a recursive algorithm, one it calls itself can only have these pointers anew at each iteration, only the final value will be printed of its address, each being new at each iteration. In a way we would not get the "array" of numbers which we could use at a later part of code to sum them altogether.

The code above illustrates my attempt at logging values per factorial through STATIC pieces of memory, h, which are, by nature of being in a function, transient in DYNAMIC MEMORY/ RANDOM ACCESS MEMORY.

The solution to this issue is to add other functions which process and maintain a value which operate before the recursion. I think, though, that this will suffer the same untimely end by virtue of a RECURSIVE PROPERTY.